organize components into subfolders (fixed some import statements) (still filled with errors)
