+++ components/shared/PasswordGate.tsx
